requirements.txt
fastapi
uvicorn
supabase
python-dotenv
openai
pydantic
pydantic-settings
websockets
pytest
pytest-asyncio
httpx
